# lesson2/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "June-Okten",
    "version": "1.0",
    "contact": {
      "email": "shirpalalona@gmial.com",
      "name": "Alona Shirpal"
    }
  },
  "paths": {
    "/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "operationId": "getAllUsers",
        "description": "Gets users",
        "parameters": [
          {
            "name": "perPage",
            "in": "query",
            "default": 20,
            "type": "number",
            "description": "Number of users per page"
          },
          {
            "name": "page",
            "in": "query",
            "default": 1,
            "type": "number",
            "description": "Page number"
          },
          {
            "name": "name",
            "in": "query",
            "default": "Climentina",
            "required": true,
            "type": "string",
            "description": "User`s name"
          },
          {
            "name": "email",
            "in": "query",
            "default": "example@gmail.com",
            "required": true,
            "type": "string",
            "description": "User's email "
          },
          {
            "name": "role",
            "in": "query",
            "enum": [
              "user",
              "admin"
            ],
            "default": "user",
            "required": false,
            "type": "array",
            "description": "User's role "
          }
        ],
        "responses": {
          "200": {
            "description": "Users found",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/singleUser"
              }
            }
          },
          "400": {
            "description": "Some error",
            "schema": {
              "properties": {
                "msg": {
                  "type": "string",
                  "example": "Something went wrong"
                },
                "code": {
                  "type": "number",
                  "example": 400
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Users"
        ],
        "operationId": "createUser",
        "description": "Create user",
        "parameters": [
          {
            "name": "name",
            "in": "formData",
            "default": "Climentina",
            "required": true,
            "type": "string"
          },
          {
            "name": "email",
            "in": "formData",
            "default": "example@gmail.com",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "formData",
            "default": "Test$100",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "User created",
            "schema": {
              "type": "object",
              "properties": {
                "_id": {
                  "type": "string",
                  "example": "a234nfj1234",
                  "description": "MongoDB ID"
                },
                "email": {
                  "type": "string",
                  "example": "Example@gmail.com"
                },
                "name": {
                  "type": "string",
                  "example": "Climentina"
                }
              }
            }
          },
          "400": {
            "description": "Some error",
            "schema": {
              "$ref": "#/definitions/commonError"
            }
          }
        }
      }
    },
    "/users/{user_id}": {
      "get": {
        "tags": [
          "Users"
        ],
        "operationId": "getUser",
        "description": "Finds single user by Id",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "type": "string",
            "required": true,
            "description": "User's id"
          }
        ],
        "responses": {
          "200": {
            "description": "User found",
            "schema": {
              "items": {
                "$ref": "#/definitions/singleUser"
              }
            }
          },
          "400": {
            "description": "Some error",
            "schema": {
              "$ref": "#/definitions/commonError"
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Users"
        ],
        "operationId": "UpdateUser",
        "description": "Updates user",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "type": "string",
            "required": true,
            "description": "User's id"
          },
          {
            "name": "name",
            "in": "formData",
            "type": "string",
            "required": true,
            "default": "Climentina",
            "description": "User's name"
          },
          {
            "name": "email",
            "in": "formData",
            "type": "string",
            "required": true,
            "default": "example@gmail.com",
            "description": "User's email"
          },
          {
            "name": "Authorization",
            "in": "header",
            "type": "string",
            "required": true,
            "description": "access_token"
          }
        ],
        "responses": {
          "200": {
            "description": "Updates user",
            "schema": {
              "items": {
                "$ref": "#/definitions/singleUser"
              }
            }
          },
          "400": {
            "description": "Some error",
            "schema": {
              "$ref": "#/definitions/commonError"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "operationId": "deleteUser",
        "description": "Removes single user",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "type": "string",
            "required": true,
            "description": "User's id"
          },
          {
            "name": "Authorization",
            "in": "header",
            "type": "string",
            "required": true,
            "description": "access_token"
          }
        ],
        "responses": {
          "200": {
            "description": "Deletes user",
            "schema": {
              "items": {
                "$ref": "#/definitions/singleUser"
              }
            }
          },
          "400": {
            "description": "Some error",
            "schema": {
              "$ref": "#/definitions/commonError"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "singleUser": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string",
          "example": "a234nfj1234",
          "description": "MongoDB ID"
        },
        "email": {
          "type": "string",
          "example": "Example@gmail.com"
        },
        "name": {
          "type": "string",
          "example": "Climentina"
        }
      }
    },
    "commonError": {
      "type": "object",
      "properties": {
        "msg": {
          "type": "string",
          "example": "Something went wrong"
        },
        "code": {
          "type": "number",
          "example": 400
        }
      }
    }
  }
}